requirements.txt.txt
flask
opencv-python
numpy
ultralytics
cvzone
sort-python